Major: 0.3.0.27
Bug Fixed:
1. Fixed #7 strange bug. This bug is really really interesting.
This is a bug is started from the commit
c169b2eccd181b736366fca66b35dfb7088ce75c.

At first, I think KCTextPanelManager should repaint all the KCTextPanel
at the time it repaint itself. But this cost a high CPU cost. I don't
realize what's happened.
Then I solved that bug by repaint all the panel manual. But I don't
notice that the difference between these two ways. The only different
is: All the panels is NOT repaint at time the panel manager repaint.
And the #7 mark panel bug is also cost by not update the data promptly.
The mark panel will only update at the time the block count changed from
9 to 10, or from 10 to 9. And the first block can use at every time. So
here's the point: There's a important thing we don't do is update the
panel.
There's a brand new way to update all the new panels and it's really
swift. And we don't need to use a loop this time: use a signal to notice
the panel to update. This signal will be connect at time when the panel
add to the manager. And the signal will be emit at the time
updateAllPanels() called in the panel manager. This slot will be called
at time KCTextEditor update.
All the things has been ended at here.